[CLOUD-124] Refactor tf_test (#7)
In pkg/loader/tf_test, we had some golden tests that check the entire resource
view.  We only had this for HCL, not other loader formats.  The other formats
were tested in Rego.

As we move the loaders and resource views towards Go, I would like to use
these testing facilities for those as well.

This PR basically moves:

    pkg/loader/tf_test/xyz

To:

    pkg/loader/golden_test/tf/xyz

And abstracts it so we have a place to drop other resource views we want to
test, e.g.

    pkg/loader/golden_test/cfn/xyz