[CLOUD-124] Port cloudformation resource view from Rego to Go (#9)
This ports the CFN input processing from Regula (in Rego) to Go. It is more or less a literal port of the Rego code, and it's clear that it can be much improved in the future, but it should be at feature parity with Regula for now.

Due to these changes, I also moved the intrinsics test over to the golden tests, along with all the tests that we had in the Rego code (except for the tags ones).