[CLOUD-191] query() builtin with custom resolvers
Add a new rego builtin, `snyk.query()` which allows resources to be
requested by type and also "scope", an arbitrary set of key-value pairs.
This query scope will be compared to the input scope, which is set by
the loaders, and will contain metadata about the origin of the input
(e.g. cloud region/account, or IaC filename).

Library users can inject a custom resolver chain into EngineOptions,
which will be called as fallbacks when the query scope does not match
the input scope, allowing resources to be fetched from places other than
the input.