[CLOUD-200] Multiple provider version support
# What does this do?

We expose a new function that can be used in rules to check terraform provider
version constraints.  **Note that we will (currently) rarely have an exact
version of the provider!**  This makes sense; a different version may be used
locally compared to CI.  If we just get the HCL files, we only see this block:

```
terraform {
  required_providers {
    aws = {
      version = "~> 4.0.0"
    }
  }
}
```

The plan is similar in terms of information it gives us.  As such, we need to
operate on constraints, not on exact versions.

In a rule, you can do the following:

```
import data.snyk.terraform

...

is_encrypted(bucket) {
        terraform.resource_provider_version_constraint(bucket, ">=4")
        ec := encryption_configs[_]
        to_bucket_id[ec.bucket] == bucket.id
}
```

This will check if the version constraints implemented in the HCL/plan can
be used with a provider with the given constraints.  If there is no version
constraint information in the HCL/plan, this function will always return `true`
(since the provider _may_ be a version 4.x.y in that case).

# How does this work?

We load provider version constraints from HCL files as well as tfplan files, and
make these accessible in `resource._meta.terraform.provider_version_constraint`.
This will contain a string describing the constraints, e.g.
`>1.2.3, <2, != 1.2.5`.  This format is specified here:
<https://www.terraform.io/language/expressions/version-constraints>.

We build two sets of constraints (those passed to the
`resource_provider_version_constraint` function, and those in the input) and
check if there are any versions that could possibly match both of these.
This is done by a `semver_constraints_intersect` function in Rego.  This
function is somewhat complex but I used `opa test -c` to ensure we have
100% test coverage for this file.

# Other changes

 -  Refactor the `tf` and `tfplan` input loaders to work on
    `models.ResourceState` rather than the less-typed `map[string]interface{}`.
 -  Fix package names for files containing `-` in upe fixture
 -  Separate out the `PureRegoProvider` into:
     *  `PureRegoBuiltinsProvider`, which provides alternative rego-only
        implementations for the builtins.  This is only used for `repl`/`test`.
     *  `PureRegoLibProvider`, which contains the Rego part of the API that
        should always be loaded.
 -  Run `opa test` in CI.