Fix file path location of Rego files
We were getting the location of Rego files wrong in LocalProvider, which was
trimming the prefix of all paths, resulting in paths like
`/rules/SNYK_00107/terraform_test.rego` when specifying `-d rego`.

This was not a huge issue since these Rego file paths are only used in the
test output.  However, it blocks snapshot testing, which requires the correct
locations.

I added an extra argument to the internal `parser` type to support this.

Furthermore, I moved the code that actually uses these prefixes from
`engine` to `data`, since it belongs there; we always want to load files
in a consistent manner rather than pushing it to the consumer of the
`data` package.