Fix issues with identical IDs for tfstate resources (#61)
The IaC team found some issues with overlapping IDs for resources in the tfstate
input.  This PR checks in the offending case, and modifies the helper we were
using to use plain array, which should prevent this going forward (we could use
a ID+namespace+type combinartion if we ever need to, but an array is simpler).