Add check for write consistency in snapshot testing (#72)
One footgun that we hand to the users is the ability to specify output
filenames for snapshot testing.  If they have code that looks like:

    snapshot_testing.match(["foo"], "snapshot.json")
    snapshot_testing.match(["bar"], "snapshot.json")

we will overwrite the file **twice** when `--update-snapshots` is given,
and the second call will still fail when running the tests.  This is why we
save the content of the values we wrote, so we can detect these inconsistent
writes and display a nice diff.

Also adds a trailing newline to snapshots we write to disk.