query(): ResourcesResolver fallthrough on empty
When ResourcesResolvers are chained together with `Or`, fallthrough when
the first resolver returns an empty array of resources, even when the
query scope matches the input scope.

This will allow us to write rules such as:

```
deny[info] {
	ebs_volume := ebs_volumes[_]
	volume_encryption_defaults = snyk.query({
		"resource_type": "aws_ebs_encryption_by_default",
		"scope": {
			"cloud": "aws",
			"region": ebs_volume.provider.aws.region,
		},
	})
	count(volume_encryption_defaults) > 0
	volume_encryption_defaults[0].enabled == true
	ebs_volume.encrypted == false
	info := {
		"message": "EBS Volume is not encrypted",
		"resource": ebs_volume,
	}
}
```

That work in IaC, cloud, and "IaC cloud context" scans. In a pure IaC
scan, the config loader could parse statically configured regions from
terraform providers and place them on resources - or use some abstract
placeholder value when no static region is configured, to indicate that
all resources that share that provider, share an AWS region. If an
aws_ebs_encryption_by_default is configured in the same region as
volumes, it can be taken into account.

In a cloud scan, as long as the config loaders insert cloud metadata,
this should also work.

In an "IaC cloud context" scan, when no aws_ebs_encryption_by_default is
configured, the fallthrough behavior introduced by this commit will
ensure that an AWS ResourcesResolver can retrieve the region-specific
default from the cloud.