[CLOUD-574] Add provider config metadata to tfplan and tf input loaders (#94)
This adds the provider configuration to the resource metadata.  The region, if
present, is duplicated to `meta.region`.  For example:

```json
"meta": {
  "region": "us-west-1",
  "terraform": {
    "provider_config": {
      "alias": "west",
      "region": "us-west-1"
    },
    "provider_version_constraint": "~\u003e 4.0.0"
  }
}
```

This will allow us to easily get the region.

This currently only works for `tf` and `tfplan` loaders -- unfortunately there is no
provider configuration available in `.tfstate` files.