[CLOUD-720] Support data resources in tfstate & tfplan loaders (#115)
For the tfstate loader, this only required adjusting the resource type from e.g.
`aws_iam_policy_document` to `data.aws_iam_policy_document`.

For the tfplan loader it's a bit more complex.  While these resources are missing
from `planned_values`, when I was looking at the plan produced by terraform,
I noticed they are included in `prior_state`.  So I added processing this in the
code where we visit modules and resources -- it turned out as a pretty natural
extension since the format is the same as `planned_values`.