[CLOUD-731] Convert legacy iac k8s messages to OPA style paths
The snyk legacy rules produce messages that sometimes refer to attributes (e.g.
name) rather than indices.  This is different from the stricter OPA paths we
use in Policy Engine (ints for array indices and strings for object keys).  An
example of a k8s manifest where this applies is:

    # https://kubernetes.io/docs/tasks/configure-pod-container/security-context/
    apiVersion: v1
    kind: Pod
    metadata:
      name: security-context-demo-4
    spec:
      containers:
      - name: sec-ctx-4
        image: gcr.io/google-samples/node-hello:1.0
        securityContext:
          capabilities:
            add: ["NET_ADMIN", "SYS_TIME"]

The legacy IaC rules will produce a path like:

    spec.containers[sec-snyk-ctx4].image

What we want to store going forward is the standardized:

    spec.containers[0].image