feat: use schemas to scrub sensitive fields
We already had input schemas for CloudFormation under `pkg/cfn/schemas`.
I moved these into `pkg/input/schemas/cfn` to make room for
`pkg/input/schemas/tf`.

These two packages share a parent, in which we can define the common schema.
This schema currently only did type conversions (e.g. `"true"` to `true`),
but it was easy to extend this to know about sensitive info.

The meat of this PR was figuring out how to reliably generate the schemas
for different providers.  I've documented this in a README in the
`pkg/input/schemas` package.

The scrubbing is applied to `tfplan` and `tf` inputs.