feat: improve count support (CLOUD-1090) (#152)
This PR improves our support for count and makes it easy to support similar
features like for_each as well.

Rather than producing one "template" resource if a resource has count set to a
non-zero value, we produce exactly the right amount of resources in a list,
which is much closer to what terraform does.

This required a surprising amount of changes.  The biggest roadblock was the
way we treated everything like a tree of expressions.  This meant we would
have something like the following expressions:

	aws_s3_bucket.counted[*].acl
	aws_s3_bucket.counted[*].bucket_prefix

Which represented many "small" objects (e.g. `{"acl": "private"}`) which had
to be re-asssembled into bigger ones, which turned out to be sort of a mess
with adding wildcards in `FullName`s and then filling them in for the `[*]`
parts.

Instead, I refactored the way we treat expressions to use Terms instead --
single expressions or collections of expressions (like resources, for example).
These terms can have a count set on them, and that way we can simply evaluate
them multiple times if appropriate.

This refactoring allowed me to remove and improve some other code:

 -  ValTree: due to the fact that we don't have many small expressions, this
    sparse/mutable ValTree optimization is no longer needed and we can use
    `cty.Value` directly.  This goes together with a change where we make less
    copies of the tree in `hcl_interpreter.go`

 -  LocalName: we can get rid of the int-based indices since terms will always
    have string-based names.  This turns `[]interface{}` into `[]string` which
    removes a bunch of casts everywhere.

 -  Phantom attributes: this is a more explicit implementation of the way we
    add extra attributes like `.arn` and `.bucket`.  We used to take care of
    this rather hackily in `dependencies()` in `hcl_interpreter.go`.