fix: panic using unset vars in locals (CLOUD-1200)
This test case highlighted a couple of issues.  I've grouped them per file:

1.  `hcl_interpreter.go`: when dealing with variable dependencies, we were
    concluding that a variable can be used whenever its not a module input.
    This is not true: it may not have a default, so we must check that it
    exists first using `LookupByPrefix`.

2.  `names.go`: refuse cases in `AsResourceName()` that are obviously not
    resources.  This was not a problem before since, we were usually testing
    `AsVariable()` before calling `AsResourceName()`, but we shouldn't rely
    on that behaviour.

3.  `phantom_attrs.go`: check for null before patching.

4.  I also renamed the test cases in `pkg/input/golden_test` slightly, moving
    them from `issue-XYZ` to `regula-issue-XYZ` where appropriate.

(1) and (2) avoided the panic, and actually improved the output, since we
were evaluating any term that had `null` variables to `null` before (e.g.
`locals.tags` in the attached test), and we can now splice in strings such as
`var.service`, so we actually evaluate the tags.  You can also see this improve
the `vars.json` golden test.

(3) is actually what caused the panic.  Even though this is not necessary for
the specific test with (1) and (2) in place, it seems sensible to include it.