fix: panic on null for_each value (CLOUD-1215) (#174)
Part of the data resource in the `for_each` in the attached test case can not
be evaluated.  This results in a `Null` forEachVal.

We were checking this for `IsKnown()` but we should be testing `IsNull` as well.

If this happens, we completely ignore the `for_each` and treat it like a "normal" resource.