fix: deny parsing for key-less rules (CLOUD-1433) (#199)
In #193 we added some code to make `query()` return different queries based on
whether or not the underlying rule has a key.

For example, it would return:

-   the query `deny` for the rule `deny { false }`
-   the query `deny[_]` for the rule `deny[info] { info = "foo" }`

However, this is inherently flawed since it is purely based on the syntax of
the underlying rule, and not its semantics.  For examples, these two shapes of
rules should be handled the same way:

-   `deny[info] { ... }`
-   `deny = library.make_deny(...)`

The code interpreting the result _already knows_ if it's expecting a set or
a single result, and that's what the query should be based on -- not the syntax
of the rule.

To fix this, I added two queries, `query` and `queryElem`.

---

When researching other uses of `hasKey()`, I found that we are also using this
incorrectly to classify rules in the factory code.  The appropriate thing to do
there seems to use OPA's typing modules to look at the _type_ of the judgement
rule rather than the _syntax_.  That's out of scope for this PR though, since
it seems that was already an issue before the introduction of #193.