fix: CLOUD-1530 remove support for old controls format
This PR removes the code path that supported the old controls format. We
only needed this backwards compatibility temporarily when we were migrating
to the new controls format. This also fixes a bug where some newly-added
rule result fields were not added to the custom unmarshal function and were
missing in unmarshalled results.