fix: CLOUD-1495 missing rule result fields
This commit fixes a bug where rule results that are unmarshalled using the
model classes that we provide were missing the Kind and RuleBundle fields.

It also fixes a logic bug that caused Kind to be missing when no metadata
rule was defined.

The unmarshal bug was quite tricky, because new metadata and rule result
fields have to be added in a few places. We should follow up with a longer
term fix that removes the backwards compatibility code, because it's no
longer needed. That means that we'll also need to go through every consumer
to ensure that all test data has been updated.