chore: track resource types outside of input resolver
The `inputResolver` was responsible for two things:

- Returning resources from the input
- Tracking which resource types were used by policy

This change splits that apart, adding a new `trackResourceTypes` to `Query`
takes care of the latter.

Aside from making input resolver code a bit simpler, this has the advantage
that we can now create the `Query` once (per input) and still set the tracked
resource types map per policy.  This will allow good caching in `Query`.

It would be better to make a caching `ResourcesResolver`, but unfortunately
it doesn't have the right return type (and asking users of the library to
produce a Rego `ast.Term` is weird...).  This results in the unfortunate fact
that we need to change the thing we pass around from a `ResourcesResolver`
to a `Query` everywhere.

Since we are changing the types around this a bit; I also renamed `policy.Query`
to `policy.ResourcesQueryCache`, since query is very much and overloaded word.