chore: remove __resources_by_type (#248)
Quite a while ago we added `__query` as an enhancened version of
`__resources_by_type` that can take "scope" into account.  This would allow us
to retrieve runtime resources using cloud context.

`__resources_by_type` can be written in terms of `__query`, so we can get rid
of a builtin by doing this.