Fix non-string variables defaulting to strings (CLOUD-1717)
Currently, unset variables default to their variable name.  This is a somewhat
unfortunate artefact inherited from Regula, and may be worth fixing at a
different point.  But it is somewhat useful if variable names are used as
resource IDs; e.g.:

    variable "app" {
    	type = string
    }

    resource "aws_s3_bucket" "main" {
    	bucket_prefix = "store-${var.app}"
    }

And then you'd see `"store-var.app"` in the output which makes some rules easier
to write (which is debatable).

This changes that behavior for **non-string** variables.  Since we know these
aren't strings, setting them to the variable name breaks evaluation by throwing
type errors in many cases, for example:

    variable "tags_1" {
    	type = map
    }

    variable "tags_2" {
    	type = map
    	default = {"foo": "bar"}
    }

    resource "aws_s3_bucket" "main" {
    	tags = merge(var.tags_1, var.tags_2)
    }

...where we try to merge an object with the string `"var.tags_1"`.

Terraform sets unset default variables to `null`, so we should follow that:
https://developer.hashicorp.com/terraform/language/values/variables#disallowing-null-input-values

A follow-up change to this could include setting zero values (e.g. an empty
list for lists) when `nullable = false`.