fix: remote submodule register bug (#258)
policy-engine does not download remote submodules by itself.  If one is used,
e.g.:

    module "remotemodule" {
      source = "terraform-aws-modules/aws/lambda"
    }

We expect the user to run `terraform init` to download these modules (as some of
them may be private and require creds).

`terraform init` will download these modules to somewhere in `.terraform` and
update the file `.terraform/modules/modules.json`.  This second file is referred
to as the "module register" in policy-engine code.

An example of this file would be:

    {
      "Modules": [
        ...
        {
          "Key": "localmodule.remotemodule",
          "Source": "registry.terraform.io/terraform-aws-modules/lambda/aws",
          "Version": "4.18.0",
          "Dir": ".terraform/modules/localmodule.remotemodule"
        }
        ...
      ]
    }

The policy-engine will read this file that `terraform init` produces, and
through that we can infer that the sources for this remote module can be found
in `.terraform/modules/localmodule.remotemodule`.

Terraform behaviour slightly changed during the time we introduced this --
you can see that it prepended a domain name (`registry.terraform.io`) to the
module `source`.  This is problematic because we used to use `source` as a
key to find the right module in the register.

---

This PR changes that to use the `Key` instead, which is expected to be more
stable.

The key is the same as the module name, which we already computed in other
places.  However, in order to do that, we need to compute this module name
**earlier** so we have it available when we read the module register.

This is why add an extra parameter to `ParseFiles` so it can be constructed
right at the start, rather than doing it later during the `VisitModule` pass.
Since we compute it earlier, I also stored it in the `ModuleMeta` so we can
remove it as an extra parameter to `VisitModule`.

We also add a test which reproduces all of this.