fix: spurious warnings for default for_each iterator
The policy engine is generating warnings indicated that `each` may be a missing
term.  This should not happen, we have some code to filter that out here:

https://github.com/snyk/policy-engine/blob/549c2fab2214c0e6f130d355ee3b8fa7d835b552/pkg/hcl_interpreter/term.go#L159

This isn’t triggering because `iterator` is set to the default value (`"each"`),
but we represent that as `""`.  This code refactors that file to make `iterator`
non-optional.

This changes one regression related to Regula.  This regression was only about
the code producing a runtime panic so we don't change that behaviour, and the
new `null` value is more correct.