fix: store resources with a for_each as objects rather than tuples (#267)
This changes the behavior of for_each on resources. We were representing these as tuples (i.e. arrays) of resources, whereas they should be objects according to the terraform implementation. This fixes cases where HCL code would index these resources using an object key, which was yielding a type error.